+++ sdk/nodejs/util.ts
@@ -12,10 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-import * as azure from ".";
-import * as pulumi from "@pulumi/pulumi";
-import * as azurestorage from "azure-storage";
-
 type Diff<T extends string | number | symbol, U extends string | number | symbol> =
   ({ [P in T]: P } & { [P in U]: never } & { [x: string]: never })[T];
 
@@ -23,33 +19,3 @@ type Diff<T extends string | number | symbol, U extends string | number | symbol
 // with properties of the same name, but with entirely different types.
 /** @internal */
 export type Overwrite<T, U> = Pick<T, Diff<keyof T, keyof U>> & U;
-
-/** @internal */
-export function signedBlobReadUrl(
-    blob: azure.storage.Blob | azure.storage.ZipBlob,
-    account: azure.storage.Account,
-    container: azure.storage.Container): pulumi.Output<string> {
-
-    // Choose a fixed, far-future expiration date for signed blob URLs. The shared access signature
-    // (SAS) we generate for the Azure storage blob must remain valid for as long as the Function
-    // App is deployed, since new instances will download the code on startup. By using a fixed
-    // date, rather than (e.g.) "today plus ten years", the signing operation is idempotent.
-    const signatureExpiration = new Date(2100, 1);
-
-    return pulumi.all([account.primaryConnectionString, container.name, blob.name]).apply(
-        ([connectionString, containerName, blobName]) => {
-            const blobService = new azurestorage.BlobService(connectionString);
-            const signature = blobService.generateSharedAccessSignature(
-                containerName,
-                blobName,
-                {
-                    AccessPolicy: {
-                        Expiry: signatureExpiration,
-                        Permissions: azurestorage.BlobUtilities.SharedAccessPermissions.READ,
-                    },
-                },
-            );
-
-            return blobService.getUrl(containerName, blobName, signature);
-        });
-}
